test(attachments): e2e coverage for upload, hydration, and path-prefix guard
Verifies the end-to-end flow that was previously broken:
- POST /api/v1/uploads/image: auth required, MIME whitelist,
  10MB cap, returns date-stamped URL under /static/uploads/intake/.
- POST /api/v1/clothing-intakes + image_urls persists Attachment
  rows and echoes them on the create response.
- /clothing-intakes/mine and /clothing-intakes (admin) bulk-load
  the attachments and surface image_urls.
- Same shape for /api/v1/after-sales (create, /mine, admin list,
  review).
- URLs that don't start with /static/uploads/ are silently dropped
  (no path-traversal, no javascript: scheme).
- Attachment rows for an intake and an after-sale ticket are
  isolated by owner_type/owner_id.

Drives two real bug fixes uncovered while writing the suite:
1. Attachment.id and size_bytes were BigInteger, which SQLite
   (test DB) can't autoincrement — switched to Integer to match
   the rest of the project's id columns.
2. clothing_intakes.create echoed body.image_urls directly in the
   response, bypassing the path-prefix guard. Replaced with a
   bulk-load of the persisted Attachment rows so the response
   reflects what was actually stored.

+++ backend/alembic/versions/u2v3w4x5y6z7_add_attachments.py
@@ -20,12 +20,12 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "attachments",
-        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
         sa.Column("owner_type", sa.String(length=40), nullable=False),
         sa.Column("owner_id", sa.Integer(), nullable=False),
         sa.Column("url", sa.String(length=500), nullable=False),
         sa.Column("mime", sa.String(length=100), nullable=False),
-        sa.Column("size_bytes", sa.BigInteger(), nullable=False),
+        sa.Column("size_bytes", sa.Integer(), nullable=False),
         sa.Column("original_name", sa.String(length=255), nullable=True),
         sa.Column(
             "uploader_user_id",
